requirements.txt
annotated-types==0.7.0
anyio==4.4.0
arrow==1.3.0
asgiref==3.8.1
binaryornot==0.4.4
certifi==2024.2.2
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cookiecutter==2.6.0
deprecation==2.1.0
Django==5.0.4
django-cors-headers==4.3.1
django-dotenv==1.4.2
django-tailwind==3.8.0
fontawesomefree==6.5.1
gotrue==2.4.2
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
Jinja2==3.1.4
markdown-it-py==3.0.0
MarkupSafe==2.1.2
mdurl==0.1.2
packaging==24.0
--only-binary :all:
Pillow==10.2.0
postgrest==0.16.4
psycopg2-binary==2.9.9
pydantic==2.7.2
pydantic_core==2.18.3
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-slugify==8.0.4
PyYAML==6.0.1
realtime==1.0.5
requests==2.31.0
rich==13.7.1
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.0
storage3==0.7.4
StrEnum==0.4.15
supabase==2.5.0
supafunc==0.4.5
text-unidecode==1.3
types-python-dateutil==2.9.0.20240316
typing_extensions==4.12.0
tzdata==2024.1
urllib3==2.2.1
websockets==12.0
gunicorn==21.2.0
python-dotenv==1.0.1
whitenoise==6.11.0
